+++ RegistroSegundoParcial/Entidades/Mantenimiento.cs
@@ -16,11 +16,11 @@ namespace RegistroSegundoParcial.Entidades
 
         public DateTime FechaProximo { get; set; }
 
-        public int SubTotal { get; set; }
+        public double SubTotal { get; set; }
 
-        public int Itbis { get; set; }
+        public double Itbis { get; set; }
 
-        public int Total { get; set; }
+        public double Total { get; set; }
 
 
         public virtual List<MantenimientoDetalle> Detalle { get; set; }
@@ -30,7 +30,7 @@ namespace RegistroSegundoParcial.Entidades
             this.Detalle = new List<MantenimientoDetalle>();
         }
 
-        public void AgregarDetalle(int Id, int MantenimientoId, int VehiculoId, int TallerId, int ArticuloId, string Articulo, int Cantidad, int Precio, int Importe)
+        public void AgregarDetalle(int Id, int MantenimientoId, int VehiculoId, int TallerId, int ArticuloId, string Articulo, double Cantidad, double Precio, double Importe)
         {
             this.Detalle.Add(new MantenimientoDetalle(Id, MantenimientoId, VehiculoId, TallerId, ArticuloId, Articulo, Cantidad, Precio, Importe));
         }
